🚧♻️ same as last commit fixing game server to be good for udp

+++ CarSumo/PacketBuffer.cpp
@@ -1,11 +1,11 @@
 #include "udpPCH.hpp"
 
-size_t PacketBuffer::Send(const UDPSocketPtr& socket, const Packet& packet, const SocketAddress& address)
+int PacketBuffer::Send(const UDPSocketPtr& socket, const Packet& packet, const SocketAddress& address)
 {
 	return socket->SendTo(packet.GetData(), packet.GetSize(), address);
 }
 
-size_t PacketBuffer::Receive(const UDPSocketPtr& socket, Packet& packet, SocketAddress& address)
+int PacketBuffer::Receive(const UDPSocketPtr& socket, Packet& packet, SocketAddress& address)
 {
 	std::vector<uint8_t> buffer(MAX_PACKET_SIZE);
 	int bytesReceived = socket->ReceiveFrom(buffer.data(), MAX_PACKET_SIZE, address);
